Refactor: Remove existing route files and models, and reintroduce them with updated structure and functionality
- Deleted old route files: authRoutes.js, batchRoutes.js, courseRoutes.js, leaveRoutes.js, studentEnrollmentRoutes.js, userRoutes.js
- Deleted old model files: AdminDashboard.js, Batch.js, Course.js, Leave.js, StudentEnrollment.js, User.js
- Added new route files with enhanced functionality and structure for authentication, batch management, course management, leave requests, student enrollment, and user management.
- Introduced a new AdminDashboard route for fetching dashboard statistics.
- Updated database configuration for improved connection handling.
- Refactored authentication middleware for better error handling and token verification.
