+++ backend/package.json
@@ -3,8 +3,10 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "start": "node src/index.js",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "nodemon src/index.js",
     "dev": "nodemon src/index.js",
+    "build": "echo \"Error: no build specified\" && exit 1",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "prettier": "prettier --write .",
@@ -12,6 +14,9 @@
     "format": "npm run prettier && npm run lint:fix",
     "format:check": "npm run prettier:check && npm run lint"
   },
+  "author": "",
+  "license": "ISC",
+  "description": "",
   "dependencies": {
     "bcrypt": "^5.1.1",
     "bcryptjs": "^3.0.2",
